moving stress tests to another class. ignoring for now
